day 6, part 2 - done

+++ src/main/java/y2023/day6/Race.java
@@ -8,5 +8,5 @@ import lombok.Data;
 public class Race {
 
     int time;
-    int record;
+    long record;
 }
